improve server

+++ client/src/app/shared/components/checkout-payment/checkout-payment.component.ts
@@ -1,3 +1,6 @@
+import { Observable } from 'rxjs';
+import { IUser } from 'src/app/shared/models/user';
+import { AccountService } from './../../../account/account.service';
 import { environment } from './../../../../environments/environment';
 import { ToastrService } from 'ngx-toastr';
 import { CheckoutService } from './../../../checkout/checkout.service';
@@ -8,13 +11,13 @@ import {
   Component,
   ElementRef,
   Input,
-  KeyValueDiffers,
   OnDestroy,
   ViewChild,
 } from '@angular/core';
 import { IBasket } from '../../models/basket';
 import { IOrderToCreate } from '../../models/order';
 import { Router, NavigationExtras } from '@angular/router';
+
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
@@ -41,17 +44,21 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   cardNumberHandler: stripe.elements.handler = this.onChange.bind(this);
   cardExpiryHandler: stripe.elements.handler = this.onChange.bind(this);
   cardCvcHandler: stripe.elements.handler = this.onChange.bind(this);
+  private user$: Observable<IUser> = null!;
 
   constructor(
     private basketService: BasketService,
     private checkoutService: CheckoutService,
     private toastr: ToastrService,
-    private router: Router
-  ) {}
+    private router: Router,
+    private accountService: AccountService
+  ) {
+    this.user$ = this.accountService.currentUser$;
+  }
 
   ngAfterViewInit(): void {
     this.stripe = Stripe(this.stripePublicKey);
-    var elements = this.stripe.elements({ locale: 'en'});
+    var elements = this.stripe.elements({ locale: 'en' });
 
     let classes = {
       base: 'stripe-element',
@@ -109,16 +116,48 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   async onSubmitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
     const orderToCreate = this.getOrderToCreate(basket);
+    const nameOnCard = this.checkoutForm
+      .get('paymentForm')
+      ?.get('nameOnCard')?.value;
+
+    const addressData = this.checkoutForm.get('addressForm')?.value;
+
+    const address = {
+      line1: addressData.street,
+      city: addressData.city,
+      state: addressData.state,
+      postal_code: addressData.zipCode,
+    } as stripe.BillingDetailsAddress;
+
+    const paymentData = {
+      payment_method: {
+        card: this.cardNumber,
+        billing_details: {
+          name: nameOnCard,
+          address,
+        },
+      },
+    } as stripe.ConfirmCardPaymentData;
 
     try {
       const order = await this.checkoutService.createOrder(orderToCreate);
       this.toastr.success('Order created');
-      this.basketService.deleteLocalBasket(basket.id);
-      console.log(order);
-      const navigationExtras: NavigationExtras = { state: order };
-      this.router.navigate(['checkout/success'], navigationExtras);
+
+      var response = await this.stripe.confirmCardPayment(
+        basket.clientSecret ?? '',
+        paymentData
+      );
+
+      if (response.paymentIntent) {
+        console.log(response.paymentIntent);
+        this.basketService.deleteLocalBasket(basket.id);
+        const navigationExtras: NavigationExtras = { state: order };
+        this.router.navigate(['checkout/success'], navigationExtras);
+      } else {
+        throw new Error(response.error?.message);
+      }
     } catch (error) {
-      this.toastr.error(error.message);
+      this.toastr.error(error.message, 'Payment Error');
     }
   }
